requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.23
psycopg2-binary==2.9.9   # optional on Windows – omit if using SQLite only
alembic==1.12.1

# ── Auth / Security ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Document processing ────────────────────────────────────────────────────────
python-multipart==0.0.6
python-docx==1.1.0
docx2txt==0.8
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=10.1.0

# ── AI / ML ────────────────────────────────────────────────────────────────────
openai==1.3.0
langchain==0.1.0

# ── HTTP / Utilities ───────────────────────────────────────────────────────────
httpx==0.25.2
requests==2.31.0
aiofiles>=23.2.1
python-dotenv==1.0.0
python-dateutil==2.8.2
cors==1.0.1

# ── Templating ─────────────────────────────────────────────────────────────────
jinja2==3.1.2
markdown==3.5.1
